requirements.txt
typer==0.9.0
openai==1.3.7
colorama==0.4.6
pyyaml==6.0.1
requests==2.31.0
rich==13.7.0
python-dateutil==2.8.2
